rename package to periphboard
