+++ periphboard/board.go
@@ -1,9 +1,7 @@
 //go:build linux
 
-// Package genericlinux implements a Linux-based board making heavy use of sysfs
-// (https://en.wikipedia.org/wiki/Sysfs). This does not provide a board model itself but provides
-// the underlying logic for any Linux/sysfs based board.
-package genericlinux
+// Package periphboard implements a Linux-based board which uses periph.io for GPIO pins.
+package periphboard
 
 import (
 	"context"
